# src/main/resources/api/openapi.yaml
openapi: 3.0.2
info:
  title: FG-AI4H-OCI Platform
  description: >
    'FG-AI4H Open Code Initiative Platform'
  termsOfService: https://github.com/FG-AI4H/annotation-tool/blob/master/LICENSE
  contact:
    email: ml@mllab.ai
  license:
    name: BSD 3-Clause License
    url: https://github.com/FG-AI4H/annotation-tool/blob/master/LICENSE
  version: 2023-02-20
servers:
  - url: https://annotation.ai4h.net/api/v1
    description: Production server
  - url: http://localhost:8080/api/v1
    description: Development server
...
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DatasetCatalogRequestStatusDto:
      type: string
      enum:
        - PENDING
        - APPROVED
        - REJECTED
        - DEPRECATED
...
    DataCatalogDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        provider:
          type: string
        provider_catalog_id:
          type: string
        location:
          type: string
        database_name:
          type: string
        aws_region:
          type: string
        bucket_name:
          type: string
    TableDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
    DataAccessRequestDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        request_date:
          type: string
        request_status:
          $ref: '#/components/schemas/DatasetCatalogRequestStatusDto'
        request_status_date:
          type: string
        motivation:
          type: string
        requester_id:
          type: string
          format: uuid
        requester_name:
          type: string
        data_owner_id:
          type: string
          format: uuid
        data_owner_name:
          type: string
        dataset_name:
          type: string
        dataset_id:
          type: string
          format: uuid
paths:
...
  /data_catalogs:
    get:
      tags:
        - Catalog
      summary: Get data catalogs
      description: Returns the list of adata catalogs
      operationId: getDataCatalogs
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DataCatalogDto'
        '404':
          description: Error
          content: { }
    post:
      tags:
        - Catalog
      summary: Creates a new data catalog for the given request
      description: Creates a new data catalog for the given request
      operationId: addDataCatalog
      requestBody:
        description: New Catalog Request object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataCatalogDto'
      responses:
        '201':
          description: Catalog added successfully
  /data_catalogs/{data_catalog_id}:
    get:
      tags:
        - Catalog
      summary: Get data catalog
      description: Returns data catalog corresponding to the uuid
      operationId: getDataCatalogById
      parameters:
        - name: data_catalog_id
          in: path
          description: Data Cataliog Identifier
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataCatalogDto'
        '404':
          description: Error
          content: { }
    put:
      tags:
        - Catalog
      summary: Update the data catalog
      description: Update the given data catalog
      operationId: updateDataCatalog
      parameters:
        - name: data_catalog_id
          in: path
          description: DataCatalog Identifier
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: New DataCatalog object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataCatalogDto'
      responses:
        '204':
          description: DataCatalog updated successfully
        '500':
          description: Given DataCatalog ID doesn't exist
          content: { }
    delete:
      tags:
        - Catalog
      summary: Delete a DataCatalog
      description: Delete the DataCatalog of given id
      operationId: deleteDataCatalogById
      parameters:
        - name: data_catalog_id
          in: path
          description: DataCatalog Identifier
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: successful operation
        '404':
          description: Given DataCatalog ID doesn't exist
          content: { }
  /data_catalogs/{data_catalog_id}/{table_name}:
    get:
      tags:
        - Catalog
      summary: Get table from data catalog
      description: Returns the table from data catalog corresponding to the uuid
      operationId: getTableFromDataCatalog
      parameters:
        - name: data_catalog_id
          in: path
          description: Data Cataliog Identifier
          required: true
          schema:
            type: string
            format: uuid
        - name: table_name
          in: path
          description: Table name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableDto'
        '404':
          description: Error
          content: { }
...
  /data_access_request/owner:
    get:
      tags:
        - Catalog
      summary: Get all data request access for current user
      description: Returns the list of data request access
      operationId: getOwnerDataAccessRequests
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DataAccessRequestDto'
        '404':
          description: Error
          content: { }
  /data_access_request/requester:
    get:
      tags:
        - Catalog
      summary: Get all data request access for current user
      description: Returns the list of data request access
      operationId: getRequesterDataAccessRequests
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DataAccessRequestDto'
        '404':
          description: Error
          content: { }
  /data_access_request/{data_access_request_id}:
    get:
      tags:
        - Catalog
      summary: Get data request access
      description: Returns the data request access corresponding to the uuid
      operationId: getDataAccessRequestById
      parameters:
        - name: data_access_request_id
          in: path
          description: Data Access Request Identifier
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataAccessRequestDto'
        '404':
          description: Error
          content: { }
    put:
      tags:
        - Catalog
      summary: Update the data request access
      description: Update the given data request access
      operationId: updateDataAccessRequest
      parameters:
        - name: data_access_request_id
          in: path
          description: Data request access Identifier
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: New Data request access object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataAccessRequestDto'
      responses:
        '204':
          description: Data request access updated successfully
        '500':
          description: Given data request access ID doesn't exist
          content: { }
    delete:
      tags:
        - Catalog
      summary: Delete a data request access
      description: Delete the data request access of given data_access_request_id
      operationId: deleteDataAccessRequestById
      parameters:
        - name: data_access_request_id
          in: path
          description: Data request access Identifier
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: successful operation
        '404':
          description: Given Dataset ID doesn't exist
          content: { }
  /data_access_request/{data_access_request_id}/update_status/{status}:
    get:
      tags:
        - Catalog
      summary: Update the status of the data access request
      description: Update the status of the data access request
      operationId: updateDataAccessRequestStatus
      parameters:
        - name: data_access_request_id
          in: path
          description: Data request access Identifier
          required: true
          schema:
            type: string
            format: uuid
        - name: status
          in: path
          description: Data request access status
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Data request access updated successfully